API review: move avocado.linux.software_manager to avocado.utils.software_manager
And while doing that, also change the import style on sysinfo.py so that it
matches the most common style in our tree, that is, individual utility modules
instead of the whole utils module.
