+++ avocado/utils/software_manager.py
